covid-brasil-dashboard/scripts/requirements.txt
requests==2.31.0
pandas==2.0.3
matplotlib==3.7.1
plotly==5.18.0
geopandas==0.13.2
fpdf2==2.7.7
python-dotenv==1.0.0